Merge "omap fixes for v3.15 merge window" from Tony Lindgren:

Fixes for omaps that would be good to get in before v3.15-rc1.

* tag 'omap-for-v3.15/fixes-for-merge-window' of git://git.kernel.org/pub/scm/linux/kernel/git/tmlind/linux-omap:
  ARM: OMAP4: hwmod data: correct the idlemodes for spinlock
  ARM: dts: am33xx: correcting dt node unit address for usb
  ARM: dts: omap4/5: Use l3_ick for the gpmc node
  CLK: TI: OMAP4/5/DRA7: Remove gpmc_fck from dummy clocks
  ARM: OMAP4: Fix definition of IS_PM44XX_ERRATUM
  ARM: OMAP2+: INTC: Acknowledge stuck active interrupts

+++ drivers/mmc/card/queue.c
@@ -197,7 +197,7 @@ int mmc_init_queue(struct mmc_queue *mq, struct mmc_card *card,
 	struct mmc_queue_req *mqrq_prev = &mq->mqrq[1];
 
 	if (mmc_dev(host)->dma_mask && *mmc_dev(host)->dma_mask)
-		limit = dma_max_pfn(mmc_dev(host)) << PAGE_SHIFT;
+		limit = (u64)dma_max_pfn(mmc_dev(host)) << PAGE_SHIFT;
 
 	mq->card = card;
 	mq->queue = blk_init_queue(mmc_request_fn, lock);
